最多5，否则马上返回

+++ roop/ui.py
@@ -897,8 +897,15 @@ def start_swap(enhancer, detection, keep_fps, keep_frames, skip_audio, face_dist
     is_processing = True
 
     if target_files is None or len(target_files) <= 0:
+        is_processing = False
         return gr.Button.update(variant="primary"), None, None
     
+
+    if len(target_files) > 5:
+        is_processing = False
+        gr.Info(f"You can upload up to 5 files at a time. If you have more needs, please contact the blogger.")
+        return gr.Button.update(variant="primary"), None, None
+
     if roop.globals.CFG.clear_output:
         shutil.rmtree(roop.globals.output_path)
 
@@ -1022,7 +1029,6 @@ def on_destfiles_changed(destfiles):
 
     if len(destfiles) > 5:
         gr.Info(f"You can upload up to 5 files at a time. If you have more needs, please contact the blogger.")
-        return gr.Slider.update(value=0, maximum=0)
 
     nsfw_detected_and_removed = False
     for file_obj in destfiles:
